Add new warnings to the c++ compile command
and fix some warnings which were reported by the new flags.
I do not consider this list of warnings final: warnings can be removed from the command line if
they prove to be a nuissance for some programming patterns. Also, new warnings can be added if
the existing codebase compiles cleanly with them.